Refactor to FriendlyPromotion::FriendlyPromotionAdjuster
This changes the API of the main class we're working with from
`SolidusFriendlyPromotions::OrderDiscounter` to
`SolidusFriendlyPromotions::FriendlyPromotionsAdjuster`. This class now
simply calls the three other (otherwise somewhat confusingly named)
service objects:

- `LoadPromotions`
- `DiscountOrder`
- `PersistDiscountedOrder`

This makes the code easier to read in my opinion, and simplifies the
API for promotion handlers significantly.
