+++ spec/models/solidus_friendly_promotions/friendly_promotion_adjuster/discount_order_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-RSpec.describe SolidusFriendlyPromotions::FriendlyPromotionDiscounter do
+RSpec.describe SolidusFriendlyPromotions::FriendlyPromotionAdjuster::DiscountOrder do
   context "shipped orders" do
     let(:promotions) { [] }
     let(:order) { create(:order, shipment_state: "shipped") }
